Scripts/DatabankLib/requirements-dev.txt
buildh>=1.6.1
MDAnalysis>=2.7.0
numpy>=1.20.0,<2.0
pandas>=2.0.0
PyYAML>=6.0.0
tqdm>=4.5.0
periodictable>=1.5.0
Deprecated
scipy
flake8
pep8-naming
pytest
